pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm[toml]>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "catphan-analysis"
dynamic = ["version"]
description = "Validated, professional software package for CatPhan phantom DICOM analysis"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Medical Physics Team", email = "jwatkins@toh.ca"}
]
keywords = ["medical physics", "catphan", "DICOM", "QA", "phantom analysis", "radiotherapy"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.19.0",
    "scipy>=1.5.0",
    "matplotlib>=3.3.0",
    "pydicom>=2.0.0",
    "alexandria-project",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=3.0.0",
]
build = [
    "pyinstaller>=6.0",
]

[project.urls]
Homepage = "https://github.com/DunirkDynamo/XVI-CatPhan"
Documentation = "https://DunirkDynamo.github.io/XVI-CatPhan/"
Repository = "https://github.com/DunirkDynamo/XVI-CatPhan"
Issues = "https://github.com/DunirkDynamo/XVI-CatPhan/issues"

[project.scripts]
catphan-analyze = "catphan_analysis.main:main"
catphan-listen = "catphan_analysis.listen_and_analyze:main"
catphan-select = "catphan_analysis.select_and_analyze:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["catphan_analysis*"]

[tool.setuptools_scm]
version_file = "src/catphan_analysis/_version.py"
fallback_version = "0.0.0"
local_scheme = "no-local-version"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | docs
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=catphan_analysis --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# mypy configuration
[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

# Coverage configuration
[tool.coverage.run]
source = ["catphan_analysis"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
